Move language badge group service
