+++ ngx-fudis/projects/ngx-fudis/src/lib/components/language-badge-group/language-badge-group.component.ts
@@ -6,7 +6,7 @@ import {
 } from '../../types/miscellaneous';
 import { FudisTranslationService } from '../../utilities/translation/translation.service';
 import { TooltipApiDirective } from '../../directives/tooltip/tooltip-api.directive';
-import { FudisLanguageBadgeGroupService } from './language-badge-group.service';
+import { FudisLanguageBadgeGroupService } from '../../services/language-badge-group/language-badge-group.service';
 
 @Component({
 	selector: 'fudis-language-badge-group',
