+++ ngx-fudis/projects/ngx-fudis/src/public-api.ts
@@ -57,7 +57,7 @@ export { HeaderDirective } from './lib/directives/content-projection/header/head
 export { HeadingComponent } from './lib/components/typography/heading/heading.component';
 export { IconComponent } from './lib/components/icon/icon.component';
 export { LanguageBadgeGroupComponent } from './lib/components/language-badge-group/language-badge-group.component';
-export { FudisLanguageBadgeGroupService } from './lib/components/language-badge-group/language-badge-group.service';
+export { FudisLanguageBadgeGroupService } from './lib/services/language-badge-group/language-badge-group.service';
 export { LinkComponent } from './lib/components/link/link.component';
 export { NotificationsDirective } from './lib/directives/content-projection/notifications/notifications.directive';
 export { NotificationComponent } from './lib/components/notification/notification.component';
